Enhance code structure by adding client code
